feat: move resolving extractors from names into `osvscanner`
