+++ cmd/osv-scanner/scan/image/command.go
@@ -76,9 +76,9 @@ func action(_ context.Context, cmd *cli.Command, stdout, stderr io.Writer) error
 	scannerAction.IsImageArchive = cmd.Bool("archive")
 	scannerAction.ExperimentalScannerActions = helper.GetExperimentalScannerActions(cmd)
 
-	if len(scannerAction.Extractors) == 0 {
-		return errors.New("at least one extractor must be enabled")
-	}
+	// if len(scannerAction.Extractors) == 0 {
+	// 	return errors.New("at least one extractor must be enabled")
+	// }
 
 	var vulnResult models.VulnerabilityResults
 	//nolint:contextcheck // passing the context in would be a breaking change
