+++ cmd/osv-scanner/scan/source/command.go
@@ -134,9 +134,9 @@ func action(_ context.Context, cmd *cli.Command, stdout, stderr io.Writer) error
 	scannerAction.CallAnalysisStates = callAnalysisStates
 	scannerAction.ExperimentalScannerActions = experimentalScannerActions
 
-	if len(experimentalScannerActions.Extractors) == 0 {
-		return errors.New("at least one extractor must be enabled")
-	}
+	// if len(experimentalScannerActions.Extractors) == 0 {
+	// 	return errors.New("at least one extractor must be enabled")
+	// }
 
 	var vulnResult models.VulnerabilityResults
 	//nolint:contextcheck // passing the context in would be a breaking change
